+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.52.0] - 2024-04-11
+### Added
+- Added connected card to the A4A plugin. [#36747]
+- Jetpack Components: add a new prop for custom classes for the ActionButton component [#36771]
+
 ## [0.51.0] - 2024-04-08
 ### Added
 - Added AutomatticIconLogo and AutomatticForAgenciesLogo components. [#36664]
@@ -1003,6 +1008,7 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[0.52.0]: https://github.com/Automattic/jetpack-components/compare/0.51.0...0.52.0
 [0.51.0]: https://github.com/Automattic/jetpack-components/compare/0.50.5...0.51.0
 [0.50.5]: https://github.com/Automattic/jetpack-components/compare/0.50.4...0.50.5
 [0.50.4]: https://github.com/Automattic/jetpack-components/compare/0.50.3...0.50.4