+++ projects/packages/sync/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.11.1] - 2024-04-11
+### Changed
+- Internal updates.
+
 ## [2.11.0] - 2024-04-08
 ### Added
 - Added social_notes_config option to sync. [#36671]
@@ -1096,6 +1100,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Move sync to a classmapped package
 
+[2.11.1]: https://github.com/Automattic/jetpack-sync/compare/v2.11.0...v2.11.1
 [2.11.0]: https://github.com/Automattic/jetpack-sync/compare/v2.10.5...v2.11.0
 [2.10.5]: https://github.com/Automattic/jetpack-sync/compare/v2.10.4...v2.10.5
 [2.10.4]: https://github.com/Automattic/jetpack-sync/compare/v2.10.3...v2.10.4